change decimal to double in db schema

+++ src/main/resources/db/migration/V1__Create_currency_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS currency_rates(
     bank varchar(127) not null,
     code varchar(3) not null,
-    buy decimal not null,
-    sell decimal not null,
+    buy double not null,
+    sell double not null,
     primary key(bank,code)
 );
